Fix fatal error caused by the class definition being too early.
My bad.
